pyproject.toml
[tool.poetry]
name = "synr"
version = "0.1.0"
description = ""
authors = ["Jared Roesch <jroesch@octoml.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "*"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
